feat: expand ast-grep Clean Architecture rules to 14 deterministic rules
Add 9 new import/keyword-based rules for early detection of Clean
Architecture violations across all workspace crates:

Domain layer (3 new):
- clean-arch-domain-no-io-operations: Block std::fs, std::net, std::io
- clean-arch-domain-no-async: Detect async functions via regex
- clean-arch-domain-no-external-crates: Block anyhow, thiserror, log, regex

Use case layer (2 new):
- clean-arch-usecases-no-presentation: Block serde_json
- clean-arch-usecases-no-global-state: Block lazy_static, once_cell

Adapter/Handler layer (2 new):
- clean-arch-adapters-no-usecase-imports: Block use case imports
- clean-arch-handlers-no-serde-json: Hint for serde_json (acceptable)

Cross-crate dependencies (2 new):
- clean-arch-core-no-daemon-imports: Enforce dependency direction
- clean-arch-common-no-higher-crates: Protect lowest-level crate

Also fixes a real violation found: moved RestartOutput from usecases to
domain layer where other output types reside.

Reorganized existing 5 rules into rules/clean-architecture/ directory.
