+++ rules/clean-architecture/clean-arch-adapters-no-usecase-imports.yml
@@ -1,0 +1,18 @@
+id: clean-arch-adapters-no-usecase-imports
+language: rust
+severity: error
+message: "Adapters should not import use cases. Adapters translate data, handlers call use cases."
+note: |
+  Clean Architecture layer boundaries:
+  - Adapters: Convert between domain types and external formats
+  - Handlers: Coordinate use case execution
+  - Use cases: Implement application-specific business rules
+
+  If you need use case types in an adapter, you may need a handler instead.
+files:
+  - "crates/agent-tui-daemon/src/adapters/**/*.rs"
+rule:
+  kind: use_declaration
+  any:
+    - pattern: use crate::usecases::$$$ITEMS
+    - pattern: use super::super::usecases::$$$ITEMS
