+++ rules/clean-architecture/clean-arch-common-no-higher-crates.yml
@@ -1,0 +1,20 @@
+id: clean-arch-common-no-higher-crates
+language: rust
+severity: error
+message: "Common crate must not depend on higher-level crates."
+note: |
+  Workspace dependency direction (lower to higher):
+  common -> core -> terminal -> daemon -> ipc -> agent-tui
+
+  Common is the lowest-level shared crate.
+  It must not import from any other workspace crate.
+files:
+  - "crates/agent-tui-common/src/**/*.rs"
+rule:
+  kind: use_declaration
+  any:
+    - pattern: use agent_tui_core::$$$ITEMS
+    - pattern: use agent_tui_daemon::$$$ITEMS
+    - pattern: use agent_tui_terminal::$$$ITEMS
+    - pattern: use agent_tui_ipc::$$$ITEMS
+    - pattern: use agent_tui::$$$ITEMS
