+++ rules/clean-architecture/clean-arch-domain-no-async.yml
@@ -1,0 +1,13 @@
+id: clean-arch-domain-no-async
+language: rust
+severity: error
+message: "Domain functions should be synchronous. Async is an infrastructure concern."
+note: |
+  Async/await is a runtime concern belonging to infrastructure layer.
+  Domain logic should be pure and synchronous.
+  If you need async, the caller in the infrastructure layer should handle it.
+files:
+  - "crates/agent-tui-daemon/src/domain/**/*.rs"
+rule:
+  kind: function_item
+  regex: "async "
