+++ rules/clean-architecture/clean-arch-domain-no-external-crates.yml
@@ -1,0 +1,24 @@
+id: clean-arch-domain-no-external-crates
+language: rust
+severity: error
+message: "Domain layer should only use std library. External crates belong in outer layers."
+note: |
+  Domain purity: Only std, core, alloc are allowed.
+  - anyhow/thiserror -> define custom error types with std::error::Error
+  - log/tracing -> infrastructure concern
+  - regex -> use in adapter layer if needed
+  - Any crate dependency -> adapter or infrastructure layer
+files:
+  - "crates/agent-tui-daemon/src/domain/**/*.rs"
+rule:
+  kind: use_declaration
+  any:
+    - pattern: use anyhow::$$$ITEMS
+    - pattern: use anyhow
+    - pattern: use thiserror::$$$ITEMS
+    - pattern: use thiserror
+    - pattern: use log::$$$ITEMS
+    - pattern: use log
+    - pattern: use env_logger::$$$ITEMS
+    - pattern: use regex::$$$ITEMS
+    - pattern: use regex
