+++ rules/clean-architecture/clean-arch-domain-no-frameworks.yml
