+++ rules/clean-architecture/clean-arch-domain-no-io-operations.yml
@@ -1,0 +1,22 @@
+id: clean-arch-domain-no-io-operations
+language: rust
+severity: error
+message: "Domain layer must not perform I/O. Move file/network operations to infrastructure."
+note: |
+  I/O operations are infrastructure concerns. Domain should be pure business logic.
+  - File operations -> infrastructure layer
+  - Network calls -> infrastructure layer
+  - Use repository traits for data access
+files:
+  - "crates/agent-tui-daemon/src/domain/**/*.rs"
+rule:
+  kind: use_declaration
+  any:
+    - pattern: use std::fs::$$$ITEMS
+    - pattern: use std::fs
+    - pattern: use std::net::$$$ITEMS
+    - pattern: use std::net
+    - pattern: use std::io::Write
+    - pattern: use std::io::Read
+    - pattern: use std::io::BufReader
+    - pattern: use std::io::BufWriter
