+++ rules/clean-architecture/clean-arch-domain-no-outer-layers.yml
