+++ rules/clean-architecture/clean-arch-domain-no-serde-derive.yml
