+++ rules/clean-architecture/clean-arch-usecases-no-global-state.yml
@@ -1,0 +1,18 @@
+id: clean-arch-usecases-no-global-state
+language: rust
+severity: error
+message: "Use cases should not access global state. Pass dependencies via constructor."
+note: |
+  Dependency Injection: Use cases should receive all dependencies through their constructor.
+  Global state makes testing difficult and hides dependencies.
+  - Replace lazy_static with constructor injection
+  - Pass configuration as parameters
+files:
+  - "crates/agent-tui-daemon/src/usecases/**/*.rs"
+rule:
+  kind: use_declaration
+  any:
+    - pattern: use lazy_static::$$$ITEMS
+    - pattern: use lazy_static
+    - pattern: use once_cell::sync::Lazy
+    - pattern: use once_cell::$$$ITEMS
