+++ rules/clean-architecture/clean-arch-usecases-no-infrastructure.yml
