+++ rules/clean-architecture/clean-arch-usecases-no-presentation.yml
@@ -1,0 +1,15 @@
+id: clean-arch-usecases-no-presentation
+language: rust
+severity: error
+message: "Use cases should not handle JSON serialization. Move to adapter layer."
+note: |
+  Presentation formatting is an adapter concern:
+  - JSON serialization -> adapters/rpc.rs or adapters/snapshot_adapters.rs
+  - Use cases return domain types, adapters convert to wire format
+files:
+  - "crates/agent-tui-daemon/src/usecases/**/*.rs"
+rule:
+  kind: use_declaration
+  any:
+    - pattern: use serde_json::$$$ITEMS
+    - pattern: use serde_json
