+++ rules/clean-architecture/clean-arch-usecases-use-traits.yml
